Rename `order-taxes` lib to `order-values`, to indicate that it’s used for any calculated value (tax, total, etc) from an order
