+++ client/extensions/woocommerce/lib/order-values/index.js
