+++ client/extensions/woocommerce/lib/order-values/test/fixtures/order-with-coupons.js
